Add UserServiceResolver interface
